+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## __8.26.3f__
 
-- [EFM] Attack Command Animations now play again
-- 
+- **[EFM]** Attack Command Animations now play again
+- Added Cooldown to summoning Dream Eaters, it lasts 30 seconds.
+- Added Cooldown mentioned above for if the Dream Eater dies.